Use new hook
